requirements.txt
sounddevice
numpy
requests
keyboard
pystray
Pillow
websocket-client
pyperclip
